Created window with fixed size sidebar
Created a sample program that had a fixed size sidebar.  The main
window is resizeable in both directions, but the left sidebar is
only resizeable in the y direction.

This window arrangement uses the pack() function rather than the grid.

Also removed inv_rename_gui.pl because that was a mess that didn't
really work.